Fix bugs found in template code review
Verified by stress tests against brute-force references:
- LCA.cpp: rewrite (missing semicolon, undefined 'parent', wrong dfs args,
  forr(20) out-of-bounds read causing wrong answers); add mxN
- MST.cpp: cnt never incremented (connectivity check dead), kruskal now
  returns MST weight instead of discarding it
- tree-diameter.cpp: emp macro had stray parens (would not compile); add mxN
- hld.cpp: remove stray 'terr;' (would not compile)
- binary-lifing.cpp: k bits >= 30 were silently ignored; clamp k
- small-to-large.cpp: warn about stack/memory blowup on chain trees
- articulation-point.cpp: rewrite (low[v]>=dfn[v] should be dfn[u], parent
  edge not skipped, duplicate counting per child; undefined 'root')
- bridge.cpp: skip parent edge only once so parallel edges are not
  misreported as bridges
- BellmanFord.cpp: vs sized n -> n+1 (1-indexed OOB), rename sfpa->spfa
- floyd-warshall.cpp: declare loop variables, add INF usage note
- Edmond.cpp: accumulate capacities and multiplicities for duplicate
  edges; min-cut output lists each parallel edge
- fwt.cpp: off-by-one k<(1<<mxB) skipped the top butterfly layer, wrong
  answers for values >= 2^20
- suffixarray_and_lcp.cpp: clear cnt[] on entry (second sfa() call
  returned garbage)
- block_cut_tree.cpp: remove leftover err() debug call
- Joseph.cpp: duplicate function definition (CE if copied whole); k>n
  branch was O(n) recursion (stack overflow), now iterative; int -> ll
- persistence-seg-tree.cpp: version-clone gave 0 for n=1 (root is leaf);
  rt[] doubled to avoid OOB when all ops are clones; lc/rc int to halve
  memory; add node-budget/multi-test notes
- Array Version of Linked List.cpp: add init() with sentinels (previously
  segfaulted without undocumented setup) and usage comment
- Exgcd.cpp: int -> ll (silent overflow on 64-bit inputs)
- findPrime.cpp: etf[1] was uninitialized
- matrixMultiple.cpp: int arithmetic overflowed and no modulo; now ll+mod
- ntt.cpp / xor_basis.cpp / shoelace.cpp / Convex_hull_trick.cpp /
  system-of-linear-eqn.cpp: document silent-failure limits (CRT bound,
  2^21 length cap, 31-bit basis, 2x area output, integer-division slope)

Rebuild codebook.pdf (20 pages) including teammate's new templates.

+++ 1 Number Theory/Exgcd.cpp
@@ -1,6 +1,6 @@
-array<int,2> exgcd(int a, int b){
-	if(b == 0) return array<int,2>{1,0};
+array<ll,2> exgcd(ll a, ll b){
+	if(b == 0) return array<ll,2>{1,0};
 	auto ans = exgcd(b, a % b);
-	return array<int,2>{ans[1], ans[0] - a / b * ans[1]};
+	return array<ll,2>{ans[1], ans[0] - a / b * ans[1]};
 }
